services/auth_service/backend/requirements.txt
Django>=5.2.1
djangorestframework==3.16.0
djangorestframework-simplejwt>=5.4.0
python-decouple==3.8
psycopg2-binary==2.9.9
django-cors-headers==4.6.0
redis>=5.0.0
django-redis>=5.4.0
psutil>=6.1.0
gunicorn>=23.0.0
whitenoise>=6.8.2
cryptography>=43.0.0
requests>=2.32.0
